=== Refactor: Shifts api requests into operations.js ===
