+++ src/components/login-form/login-form.jsx
@@ -1,7 +1,7 @@
 import React, {createRef, PureComponent} from "react";
 import PropTypes from "prop-types";
 import {connect} from "react-redux";
-import {login} from "../../store/user/actions";
+import {login} from "../../store/user/operations";
 
 class LoginForm extends PureComponent {
   constructor(props) {
