+++ src/store/user/actions.js
@@ -1,5 +1,3 @@
-import {AuthorizationStatus} from "../../const";
-
 export const ActionType = {
   REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
   SET_USER_INFO: `SET_USER_INFO`,
@@ -15,21 +13,3 @@ export const setUserData = (data) => ({
   payload: data,
 });
 
-export const checkAuth = () => (dispatch, _getState, api) => (
-  api.get(`/login`)
-    .then(() => dispatch(requireAuthorization(AuthorizationStatus.AUTH)))
-    .catch((err) => {
-      throw err;
-    })
-);
-
-export const login = ({login: email, password}) => (dispatch, _getState, api) => (
-  api.post(`/login`, {email, password})
-    .then((res) => {
-      dispatch(requireAuthorization(AuthorizationStatus.AUTH));
-      dispatch(setUserData(res.data));
-    })
-    .catch((err) => {
-      throw err;
-    })
-);
