+++ src/store/user/operations.js
@@ -1,0 +1,21 @@
+import {AuthorizationStatus} from "../../const";
+import {requireAuthorization, setUserData} from "./actions";
+
+export const checkAuth = () => (dispatch, _getState, api) => (
+  api.get(`/login`)
+    .then(() => dispatch(requireAuthorization(AuthorizationStatus.AUTH)))
+    .catch((err) => {
+      throw err;
+    })
+);
+
+export const login = ({login: email, password}) => (dispatch, _getState, api) => (
+  api.post(`/login`, {email, password})
+    .then((res) => {
+      dispatch(requireAuthorization(AuthorizationStatus.AUTH));
+      dispatch(setUserData(res.data));
+    })
+    .catch((err) => {
+      throw err;
+    })
+);
